requirements.txt
asgiref==3.3.1
certifi==2020.11.8
Django==3.1.4
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
